Add reversed time in sorted set so score would be in desc order but if we have duplicate of score we would select by time in asc order; Save pp values with floating point;

+++ Server/API/Serializable/Response/UserStatsResponse.cs
@@ -18,7 +18,7 @@ public class UserStatsResponse(UserStats stats, int globalRank, int countryRank)
 
     [JsonPropertyName("play_count")] public int PlayCount { get; set; } = stats.PlayCount;
 
-    [JsonPropertyName("pp")] public short PerformancePoints { get; set; } = stats.PerformancePoints;
+    [JsonPropertyName("pp")] public double PerformancePoints { get; set; } = stats.PerformancePoints;
 
     [JsonPropertyName("rank")] public int Rank { get; set; } = globalRank;
